First wave of large-scale subclassing overhaul
Over the past few years, I've become increasingly skeptical of the
3rd-party cSelfSubHookCallback solution for subclassing.  Manually
injecting machine code at run-time is simply problematic, and I've had
to add a bunch of manual fixes to the class to work around some
obnoxious reliability issues.  (I'm also 99% sure that the class is a
primary reason that PD hasn't worked on Wine for quite a few years,
which is an annoying side-effect.)

For a variety of reasons, I've wanted to move to the much-safer,
Wine-friendly comctl32 subclassing technique (nicely outlined here:
http://vb.mvps.org/samples/HookXP/).  The trouble with doing this is
that the transition is not exactly seamless, because not only are the
two approaches very different, but they interfere with each other and
cause user object leaks when intermixed.  (As described here:
http://stackoverflow.com/questions/1120337/setwindowsubclass-is-leaking-user-objects)
I've had to deal with these leaks before when working on tooltips, and
they can become quite severe over long sessions, so this isn't a project
where I can migrate stuff over in small bits.  It's got to be "all or
nothing."

To that end, I've finally started making the transition.  I'm not going
to sync any of my changes until all classes are properly transitioned,
and I *do not* recommend reverting to these intermediary commits unless
absolutely necessary.

First up, this class introduces the ISubclass interface and migrates
over the following subclass-heavy classes:
- pdWindowManager (which got a large-scale rewrite to match)
- pdWindowPainter
- pdWindowSize
- pdWindowSync

Core subclassing helper functions have also been added to the VB_Hacks
module.

The scope of this overhaul means that some other classes are necessarily
going to be affected; for example, pdDictionary is used to track
subclassing handles in a number of places, and I've had to modify it to
improve situations where keys are frequently added and deleted.  Other
classes may have received smaller changes to help with this project;
apologies for not splitting them into their own commits.