pdInputMouse: migrate to comctl32 subclassing
This was a big one, and because it required a bunch of changes to
pdInputMouse, I went ahead and refactored the class to match.  Function
definitions have been slimmed down, window message handling is now
separated into much smaller groups, and a number of redundant settings
have been purged.

Other files affected by this are ones that use pdInputMouse for input
handling; their AppInputTracker functions had to be updated to match the
new definition.