pdInputKeyboard: convert to comctl32 subclassing
Also, convert to standard Windows hook methods.

This was the last class that required conversion, so as of now, we are
free of the old "inject machine code at runtime" approach to
subclassing.  Yay!

Most exciting for me is that this successfully solved the bizarre
problem where PD would generate several thousand user objects right
before it shutdown.  User object count is now completely stable, and
shutdown behaves as expected (with the user object count gradually
dropping as UI elements are freed).

Note that this overhaul does create potential problems with breakpoints
in the IDE, but that's okay - it's a necessary trade-off toward
improving overall stability and performance of the project.