UI: dark theme is now the default setting
The "themes" menu has also been promoted to the "Tools" menu (instead of
the "Developers only" menu), because I think it's finally ready for
primetime.

Still to-do are failsafe resource implementations if the Theme files go
missing, and a first-run screen to make the whole language and theme
selection experience less jarring.