New Tools > Theme... menu
All theme settings are now controlled via a single "theme" dialog.  This
simplifies the UI synchronization process quite a bit, and I wanted to
build a "language and theme" dialog for first-run purposes anyway, so
this solved several problems at once.

Next up is auto-displaying the dialog to users the first time they use
the program.