Text tool: redesign layout
This fixes one of the problem tools described in #233.

Next up, doing something similar for typography.

I realize that this makes PD's text tools even more "panel happy", but
to produce a big, touch-friendly UI, there's not much getting around it.
I also like this solution because it frees up room for the future
possibility of "text entry" settings, e.g. changing the size/font of the
text entry box (separate from the text layer settings).  I haven't
settled on a good way to provide this, but it would be helpful for CJK
languages, as they are particularly hard to read at small sizes.