Preliminary rotate cursor support...
...vs using the default system "resize" cursor for everything.

The goal is to eventually migrate all canvas cursors to custom ones,
which would allow us to custom-scale cursors for users who desire larger
or smaller ones (or high-contrast ones).  PD doesn't actually use that
many custom cursors at present, but that's likely to change as more
features are implemented.

I'm not pleased with the rotation icon I've got right now, so I'll be
replacing it before actually activating it.  This however gives me a
test bed to start working on the problem.