Multipage images: improve initial load performance
FreeImage uses a very strange API for multipage images.  To work around
this in a way that doesn't require loading/unloading the entire source
image between frames, there's now a dedicated function inside the
FreeImage wrapper that caches multipage image handles, and attempts to
reuse frame data when it can.  For large (in size or frame count)
images, loading performance is now much better, and it uses less memory
too.