Effects > Stylize > Vignette: improve performance
This effect wasn't behaving properly on 32-bpp base images with
semi-transparency.  That's now been fixed, and the effect itself is now
rendered via pdCompositor which makes it more or less instantaneous on
normal photo sizes.  I've also added a few extra features that didn't
occur to me previously (but are somewhat nice to have).