Move/size tool: FINALLY make corner move/size work concurrently with rotation
*plays victorious music*

I've been trying to crack this problem for ages, and every time I've
given up because I couldn't solve the geometry involved.  Right now,
however, I *really* needed to solve this problem because I need to steal
the interaction code for the selection module (so selections can finally
support the same move/rotate paradigm that layers do).

Something about that extra motivation worked, and I've finally gotten
this stupid thing working.

The problem occurred when a layer has been rotated (remember: on-canvas
rotation is non-destructive in PD), and we subsequently resize that
layer by click-dragging a corner node with the move/size tool.

By default, PD rotates layers around their center.  (This is the most
intuitive behavior, vs the top-left corner that other software sometimes
uses.)  The problem during the above stage is that the center point of a
layer is constantly changing while you click-drag that corner node,
because the layer's size and/or position is changing.  This causes the
layer to "jump" all over as it rotates around those new rotational
center points created by the drag event.

I'd previously solved this issue by making corner click-drag events
resize the image proportionally in *all* directions.  This allowed the
center of the image to remained fixed during the drag, which solved the
geometry problem... but it also created really unintuitive corner-drag
behavior.

The solution, it turns out, is to dynamically reposition the rotational
center point during the move/size event, so that it remains fixed in
terms of absolute coordinates.  (pdLayer stores the rotational center
point already, thanks to a past commit, but it stores it as a *relative*
value - e.g. [0.5, 0.5] for the center of the image.)

Then, when the mouse is released, PD must reset the point of the
rotation to match the center of the new layer size and position, and
THEN it has to solve a second equation to ensure that the layer does not
move as a result of this latest rotational center point change.  (If we
don't solve this equation, the layer will "jump" to a new position,
thanks to its angle remaining constant but the point of rotation
changing.)

Anyway, this was all a giant pain in the ass, but it's done now, meaning
that layers finally work as they should, and the code is ready for me to
steal and apply it to pdSelection.  Yay for that.