pdGlyphCollection: fix line width calculations in edge cases...
...where the current line cannot be broken "correctly" because there are
no valid line break positions (e.g. for English, a single word extends
beyond the full width of the current bounding box).

Previously, this caused a glitch in the line width renderer.  (Line
widths would not be calculated correctly because it didn't know which
line to stick the "can't git" glyphs on).  Now, however, the renderer
checks for this special case, and will follow the same convention as
other editors.  (Specifically, the line will be forcibly broken
mid-word, and all affected line widths will be re-calculated to match.)