Continued code restructuring to prepare for...
...new selection features.

I'm almost ready with a much-improved selection mask rendering pipeline,
one that will finally be performant enough to integrate selections into
the paint engine.  But before I tackle that extremely messy project, I
want to wrap up a few other side projects that have been lingering too
long.  So as it is, this commit has the (start of the) new pipeline
commented out, and the old pipeline is still used for selection
rendering.