PNG export: finally solve pesky reflow issue in export dialog
One downside of using the API for all window management is that some
window functions are "asynchronous-ish", because they rely on a series
of window messages to fully implement your request.  Manually yielding
for the message pump is sometimes required to let everything "catch up".

Also included in this commit are updated language files (as of build
2542).  Many thanks to Roy K for his ongoing work in keeping nightly
builds up-to-date!