Selection tools: new UI rendering pipeline
This has been an ugly project, but (I think?) I'm happy with the final
result.

Selection tools now use a rendering pipeline similar to PD's main
compositor, where intermediate results of the UI rendering process are
cached, and when a redraw is required, we reuse as much of the previous
pipeline pass as we can.

For example: if the underlying selection data hasn't changed, but the
user has scrolled to a new viewport position, all we need to
re-calculate is the new overlap position, and a UI mask to go with it.
We don't actually need to generate a new selection mask.

This new pipeline obviously improves selection tool responsiveness, but
that was honestly a secondary goal.  A far more important consequence is
that it allows us to skip all selection pipeline rendering while a paint
tool is active.  (This works because the selection data isn't being
changed - only the *image* data is, and with this new system, the
selection stuff is separated into its own cache, meaning we can reuse it
as-is, making it effectively require "no time" in the viewport render
chain.)

Some additional optimizations could be added in the future (for example,
"interior selections" can probably cheat and only redraw a sub-portion
of the UI overlay), but this is a sufficient improvement for now.  I
wanted the new rendering pipeline in place so that I get accurate timing
reports as I migrate the underlying selection shape engines to a new,
universal shape engine.