Image/Layer > Resize: clean up and improve code
Some of these older, foundational features are a real code-style
clusterfuck.  As part of migrating all recordable functions to XML
parameter lists, I took the chance to clean up some of the worst
offenders in PD's resize dialog.

The bad "halftone" resize parameter is also gone.  That was simply a
StretchBlt wrapper, and Halftone is just their name for a bilinear-like
average.  (Unfortunately, like most GDI functions, the actual
implementation is left up to hardware manufacturers, which mean its
behavior isn't guaranteed - not ideal for software like this.)