Fix minor bugs in "Close All" behavior
- The currently active image is now reflected in the "unsaved changes"
dialog, instead of the first image in the collection with unsaved
changes
- The tabstrip is now forcibly refreshed as images are unloaded.  This
was previously disabled for performance reasons, but the new tabstrip is
much faster, and the old behavior was really confusing because the image
being saved was rarely the image reflected as "currently selected" in
the tabstrip.
- As part of the previous step, the "scrollable" property of the image
tabstrip is now updated correctly as images are unloaded.  If "Close
All" is canceled mid-way through closing a huge list of unsaved images,
the image tabstrip will no longer show dead space beyond the end of the
list.