Magic wand selections: outline mode is now enabled
1) Magic wand selections have received a boatload of performance
improvements.  This was a necessary precursor to implementing outline
support.

2) Outline mode is no longer restricted when magic wand selections are
in effect.

3) HOWEVER, only contiguous magic wand selections currently render their
outline properly.  Global selections are another beast entirely, and I
have not yet devised a solution for them.

4) pdFloodFill and pdEdgeDetector have been updated to reflect this new
functionality.  They do most of the heavy lifting.

5) pdImage has also received some new helper functions to allow for
better compositor caching.  The magic wand tool makes use of this.

This is all somewhat preliminary, so more testing will be necessary.
The "global magic wand" problem is also a nasty one, and I don't have a
good solution yet.  Ideas welcome as performance is obviously a sticking
point here.