Fix "search and replace" helper EOF linebreaks
I sometimes use a homebrew app to perform project-wide search and
replace tasks, and the tool has been erroneously adding extra blank
lines to the end of various code files.  This adds noise to commits as
the files appear to be changed, which in turn clutters up the changelog.

This commit not only fixes the helper app's behavior, but also
standardizes EOF linebreaks across the project.