Magic wand selections: enable outline mode everywhere
There's still performance work to be done, but outlines can now be
successfully rendered around contiguous *or* global selections, with or
without "holes" (e.g. selecting a "donut" image).

The Effects > Stylize > Outline tool has also been updated to take
advantage of this new, much more advanced edge detector.

At present, *all* selection rendering modes create a backup outline
copy.  This hurts performance when the user just wants "highlight" or
"lightbox" mode behavior.  I'm working on a fix for this, but it's
somewhat involved as we need to maintain a persistent pdFloodFill
instance in the background.