Flood fill: performance and quality improvements
I'm working on a true scanline floodfill implementation (as opposed to
the simplified "test run" version in this commit), but early results are
promising.  It's the sheer volume of edge cases that make a "perfect"
result bothersome.

In the meantime, polygon rendering is now used to populate the
destination "fill DIB", which shaves precious time off the actual flood
fill (because all we do in that step is populate the fill map - which
means we're no longer bouncing around in memory due to destination DIB
accesses).  A vector approach means that we get true, "perfect"
antialiasing along wand selection edges, as opposed to a raster-based
"estimation" like the old function.  This yields meaningful performance
and quality improvements, so wins all around.