pdFloodFill: finalize scanline-based implementation
This provides a 30-40% performance boost, depending on the complexity of
the underlying area.  Final fill results are identical between the two
methods, so the only reason to use the scanline version is for
performance gains.

I'm going to permanently activate the new scanline method as of this
build, and while I don't expect any problems, please let me know if you
run into any weirdness with the magic wand tool.