pdSelection: complete large-scale performance and feature overhaul
Well, this one was a doozy.  I don't know that I've ever
performance-profiled a particular feature as intensely as this one, but
this was a critical thing to get right, as it directly impacts paint
tool and viewport performance.

Updates in this commit include...

1) PD can now generate real-time outlines for any underlying selection
mask, including "raster masks" (e.g. masks not defined by polygons, like
selections modified by Select menu commands).  This was a blocker for
"paintable" selection masks, so it's pretty awesome to have this done -
and not just done, but working efficiently.  In fact, it's so efficient
that I now use this rendering technique exclusively, even with basic
selection shapes like "rectangle" or "circle".

2) Selection rendering has been deeply perf-tested, and a bunch of new
techniques, caches, and optimizations are in play.  Magic wand
selections in particular now offer "best-in-class" performance, even
when extraordinarily complex images and/or selections are in play.  (For
example, PD now meaningfully outperforms GIMP on all magic wand modes,
and that's *despite* us rendering true vector outlines.  To see an
example of how great this is, open a large image in GIMP and PD, and
compare the results of a "circle" selection - our outline actually
traces the ellipse outline, rather than trying cheating and using BitBlt
with an offset pattern.  This makes our results far prettier on complex
selection outlines, or outlines with curved/diagonal edges.)  I've used
every trick in the book to make this possible, and though I've tested
the results intensely, I appreciate any additional crash-testing.  (A
massive amount of unsafe pointer-work is required to get this level of
performance, and I really hope I've done everything correctly...)

Also note that this latest magic wand incarnation outperforms Paint.NET
- and they rely on the (excellent) 3rd-party Clipper library for polygon
partitioning, plus they benefit from multithreading.  We are
single-threaded and doing all our own math.  (Yes, I am proud of this.
;)

3) When zoomed-in, selections now trace pixel boundaries correctly.  For
example, if you create a "circle" selection and zoom in greatly, the
curved selection outline will now accurately march around individual
pixels, as opposed to remaining curve-like.  This yields far better
results when doing precise "pixel art" type work.  (Again, because I'm
feeling sassy - this is something Paint.NET and Krita both get wrong.
GIMP does it correctly, to their credit.)

4) Marching ant outlines have received a bunch of fixes, including
disabling animations while external dialogs are active.

5) PD's central edge-detection and flood-fill classes have received
large-scale performance improvements.  This was a necessary step toward
improving magic wand performance, and now other tools can reap the
rewards.

6) I've also improved the performance of polygon selections, and
reworked their UI to better communicate polygons that are in an
"unfinished" state.

With all these new features and fixes in play, I think I'm finally ready
to tackle integration of selection masks and paint tools (so that
painting only occurs in selected areas).  This is an area that could
make-or-break PD's already tenuous hold on paint performance, but with
all these optimizations, I think we'll be able to maintain our
significant performance lead over other open-source paint engines.
Fingers crossed...

(Also, the code is finally in a state where "union selections" will be
possible, which is something I've wanted to add since selections were
first implemented!  I may not get to this in 7.0, however - but it's
absolutely on my to-do list.)

One last note: there are still some minor bugs to resolve, like marching
ant outlines incorrectly appearing along the left- and top-borders when
zoomed in.  I'm aware of the problem and should have a fix ready in the
next few days.