Rendering pipeline: track and report time in each stage
This segment of the program has received more profiling than anywhere
else, and I worry that future profiling will be like extracting blood
from a stone.  But if it proves necessary to try and eke out more
performance, it would be nice to know where to target - so debug logs
now track the amount of time spent in each pipeline stage, and write out
totals when the program is closed.