Paint tools: initial selection tool integration
After a paint stroke is released, the stroke is now masked to match the
active selection.  This is mostly a "proof of concept" implementation
that can likely be accelerated with further testing.

Still to-do is making the paint stroke "dynamically" constrained by the
active selection while the painting is occurring.  This is obviously a
much nastier problem to solve.