pdUndo: remove "auto tool select" feature
In the past, I thought it would be helpful to have the Undo/Redo engine
automatically sync the selected tool to whatever tool you were using
when a given Undo/Redo point is restored.  But in practice, I'm finding
this more annoying than helpful, as it's unexpected to have tools being
automatically changed without user input.