Paint strokes are now constrained by selections
The performance impact of painting with an active selection is extremely
low, which means real-time painting is still very much possible, even on
weak hardware.  Even better, everything should be perfectly
pixel-accurate - including painting while deeply zoomed in!

This was a primary blocker for the 7.0 release, and I am quite happy to
have it resolved.