Selections now persist between selection tool changes
The old behavior was extremely unintuitive.  Now, switching between
selection tools does not clear the current selection, bringing our
behavior into line with other editors.

This allows the user to switch between selection tools on a whim, and
the current selection is only cleared if the user actually initiates a
new selection.