Magic wand selections: fix read/write file behavior
Previously, wand selections were being treated as vector selections
defined by an initial point and threshold, but this causes serious
problems when attempting to save a selection, then load it on another
image (as the shape is lost).  This fix allows wand selections to be
saved/loaded like any other selection.