pdDIB: fix potential bug with pre-release DCs
Many functions in PD release their pdDIB object from its DC when they're
done with it, in an attempt to keep GDI object count low.  If such a DIB
is immediately re-created at the same size, PD shortcuts the process by
simply using the existing DIB.  However, the creation function also
assumed a DC is present; this is not always the case.

(The color selection dialog was raising debug warnings about this.)