Listbox support classes: add option for forcible refreshes
PD's listboxes currently paint to the screen lazily (meaning paint
commands are placed at the end of each control's window message queue).
I may want to change this to improve scrolling responsiveness, but it's
a messy change that affects quite a few related controls (e.g.
dropdowns, owner-drawn listboxes, etc), so I haven't quite decided if
it's worth the effort.

Now, at least, a framework is in place to make this possible.