Central processor: refactor in preparation for 7.0 release
The largest remaining 7.0 blocker is migrating all internal processor
functions to XML param strings.  (At present, some functions still use
the old parameter-by-ordinal approach.)

The first step toward solving this is reworking PD's central processor
to be far more streamlined and readable.  This function has grown
extremely large over the past few years, with more and more features to
cover, and it was past due for a clean-up.  (Hopefully nothing broke in
the process, hehe.)

With this done, I'm now going to start migrating functions to XML
strings in roughly menu-order.  When a whole menu's worth of functions
is fixed, their resulting parameter section is going to be removed from
the main Process() function and into a sub-function dedicated to that
menu.  This will further clean up the main processor, and also make it
easier to integrate future feature patches.

(As part of this fix, a bunch of processor-adjacent functions have been
rewritten for speed, clarity, and formatting reasons.)