Layer > Add > From visible layers
This is a common option in other editors, and it saves the user a Copy >
Paste as New Layer roundtrip.  While I was here, I also added a menu
entry for the same "new basic layer" dialog raised by the right toolbar
"add layer" button - sorry for not doing this sooner!

The march toward XML-based parameters continues, with most of the Image
and Layer menus migrated as of this commit.