Remove Effects > Light and Shadow > Lens Flare
There's a longer story here, but basically, I've just discovered that
some functions contributed to PD by an outside contributor were
plagiarized directly from other software packages.  In the case of Lens
Flare, it's a line-for-line rip of GIMP's lens flare effect, which is
GPL-licensed and thus totally incompatible with PD.

This is deeply frustrating and unexpected, but at least I'm glad I
discovered this before 7.0 released.  Unfortunately, I'm still in the
process of reviewing other contributions from this individual, and I
fear other functions may be similarly affected.

In most cases, I should be able to write my own clean-room
implementations to replace the missing functions, but lens flare is a
particularly complicated one and I don't know how to approach it.  As
such, my only option right now is to remove it, and hope that I can
develop my own implementation at some point in the future.

I apologize for the trouble.  This is not how I wanted to spend my PD
time this week.  :/  In the future, I'll be more diligent about
reviewing outside code contributions.