Effects > Nature > Freeze: rewrite as "Snow" effect instead
I like the thought of a winter-themed effect, and this is way more
practical.  The output is also pretty great, especially on wintertime
nature photos.  Cranking up the "wind" (which is basically just a motion
blur effect) can also result in a natural-looking rain effect.

Thanks also to Roy K for keeping many of PhotoDemon's machine-translated
language files up to date, and for hand-updating the German language
file for just about every nightly build.  He's amazing!