Effects > Artistic > Plastic Wrap: new effect
This effect came about the same way most effects do: I find a Photoshop
tutorial on how to do something interesting, only to get halfway through
and realize PD is lacking one of tools required by the tutorial.  In
this case, that tool was a "Plastic Wrap" effect.

Comments from the file header describe the rest:

"Plastic wrap" has been available in Photoshop for well over two decades
at this point, but I've yet to see an open-source software package
reproduce the effect well.  (For example, try GIMP's analog of it, in
the Filters > Light and Shadow menu.  It's just a Script-Fu wrapper, and
it's terrible.  Ugh.)

As usual, this means I had to design PhotoDemon's implementation from
scratch, and also as usual, I think the end result is pretty darn great!
To my knowledge, this the closest anyone's come to reproducing
Photoshop's effect, and while there are still differences (some by
design, to improve output on modern photo sizes), I think our technique
is actually the preferable one - and any differences are small enough
that you can still use our technique to mimic tutorials that utilize the
Photoshop effect.

The actual effect implementation is simple: it's basically a modified
"Metal" filter, which reduces image structure to a series of alternating
ridges and troughs, followed by a modified "Emboss" filter, which
basically calculates a slope and direction at each point in the image.
We also apply some highlight and smoothing effects, obviously, to
produce a "shiny" result that looks like the filter's namesake.

Thanks to the simplicity of our implementation, most of the work can be
done in a grayscale color space, which means performance is quite
snappy.  (In fact, I believe this is how Photoshop does it as well,
which could explain how they were able to implement a demanding filter
like this so many years ago.)

Anyway, that should explain the geometry behind our implementation, and
why the effect's settings bear some resemblance to the "Metal" and
"Emboss" effects.  I particularly like the lighting controls in our
implementation, which are a neat improvement over the Photoshop version.