Purge old parameter system code from the project
This is a huge accomplishment, and one of the final blockers on the road
to a 7.0 release.  Woohoo!

I'm going to take a quick detour the next few days to work on some bug
reports I haven't been able to crack, but once that's done, it's the
home stretch toward release.  Next up: finally solving tab order issues
once and for all.