Image > Edit Metadata now connects to the Undo/Redo engine
Metadata changes can now be reverted (or re-applied) using standard
Undo/Redo clicks.

I haven't done a ton of testing on this, alas (and probably won't before
7.0 releases), but this at least fixes #206.  It also gives me a place
to focus my attention for planned metadata enhancements in 7.2.

Thank you again to @vhreal1302 for his help with this.  I apologize that
it took me so long to fix it.