Implement Unicode menu support
This was an ugly one, but I *think* everything is working correctly.
(If it is, you should never again see ???? captions in menus, including
filenames in places like the Recent Files menu, regardless of your OS
locale settings.  Tools > Language > Chinese or Russian is an easy way
to experiment.)

PhotoDemon's menus are a complicated bit of kit, because different menus
operate in different ways.  Some are created/killed at run-time, others
are enabled/disabled under a variety of criteria, some menus use
checkboxes or run-time 32-bpp icons.  All of these various menu captions
need to support localization, obviously, including localization for
hotkeys (which are also generated at run-time).  Getting all of these to
play nicely with translations and themes was unpleasant, but I can't
exactly call PD "Unicode-ready" if it doesn't have Unicode menus, eh?

With this nasty task done, the last major non-Unicode feature in the
project is message boxes.  Message boxes are primarily used for
diagnostic and debugging purposes, but there are some that may appear to
normal users (especially during error states), so they need to be dealt
with before release.