Start work on a framework for automating tab key forwarding
The goal here is to automatically sort controls by left-to-right,
top-to-bottom order (or perhaps the opposite in RTL languages?), and
automatically forward control to the nearest visible-and-enabled control
when Tab is pressed.  Running our own UI toolkit should simplify this
greatly, although there are a number of complications with controls that
host other controls.

Anyway, I'm going to try and tackle this in steps.  First up is a basic
tracking mechanism that stores hWnd lists on a per-form basis.  We won't
actually sort controls until Tab is pressed, which means the performance
impact of this new tracking code should basically be nil.  (In fact,
dialog loading will actually be *faster* as pf this commit, because I've
disabled some old theming checks that are no longer needed.)