Finalize interfaces required for automatic OK/Cancel handling...
...via Enter/Esc keypresses.

Any dialog with a command bar at the bottom (the colored bar with OK and
Cancel buttons, and possibly preset-related buttons) is now wired up for
automatic Enter/Esc key handling.  This should make it much faster to
navigate the program via keyboard, as you won't need to navigate to the
OK/Cancel keys to trigger them - just hit "Enter" or "Esc" while any
control has focus, and "OK" or "Cancel" will be automatically triggered.

As of this commit, only pdSpinner is capable of relaying Enter key
messages to the central handler - but now that I've got working
framework, I'm about to expand support to other controls.