pdDebug: large performance improvements
Should have done this sooner!  The append log file handle is now
persistently cached for the lifetime of the class, instead of
re-creating it on every write.  (I didn't realize I had yet to change
this.)

As a result, debug logging should be far less time-consuming, especially
the large logs created during startup and image loading.