Paint tools: fix selection bounding when painting/filling vector layers
In an attempt to be helpful, if you try to paint on (or fill) a text
layer, PhotoDemon will automatically place your strokes into their own
layer "above" the text layer.  This keeps the text editable, and
prevents "accidental" rasterizing.

Unfortunately, this also requires two codepaths inside the paint engine,
and the "paint/fill on text layers" wasn't accounting for selection
masks.

Still to-do is fixing live-previews from the brush engine while the
painting actually occurs.  I'm working on that next.