Greatly improve canvas responsiveness
Profiling always turns up surprising problem locations.  A combination
of fixes manage to cut the time between mouse-release-when-painting and
"ready for next paint stroke" in half.

This commit also fixes an annoying issue where the paintbrush cursor
would disappear after a stroke is merged, and a mousemove event was
required to make it re-appear.  (One of the downsides of rendering
cursors directly onto the canvas, instead of using actual cursors, I
fear.)