Open > Recent Files: overhaul management class
This class had proven to be a surprising bottleneck for anything
involving menus, as it was traveling out the HDD on every refresh, and
loading *all* recent file thumbnails from file!  For a full Recent Files
list, that's 10 thumbnails that all had to be retrieved and decoded,
wasting a lot of time.  (Similarly, updating a single thumbnail involved
saving all thumbnails out to file, then re-loading all of them!)

I'm very grateful for the 3rd-party contributor who originally reworked
this class using a shared "recent files" interface, but because the File
> Open Recent menu has some specific needs, it was time to rewrite it as
a dedicated class.  This has made a large difference, particularly when
loading files and/or running PD from a USB drive.

The class now manages thumbnails internally.  Thumbnail files are only
loaded from HDD when PD itself is loaded.  For the remainder of the
session, the full Recent Files list (and all associated menus) are
managed in-memory.  When PD is closed, the final updated file (and
associated thumbnails) are finally written out to the HDD.