Color picker: implement core features
Still need to create a custom mouse icon for this tool, and of course
things like tool selection hotkeys still need to be handled.

But!  All other aspects of this tool now work.  Radius-based sampling is
supported, "sample merged" is available, and several color description
boxes (and an actual preview) are provided to help zero-in on the
information you want.

And with that, we're one step closer to 7.0's release...