Gradient editor: "noise" gradients are now supported
This brings PD's gradient editor to near-parity with Photoshop:

https://helpx.adobe.com/photoshop/using/gradients.html

Our current implementation is not optimal, as performance is hindered by
writing/reading all gradient points as XML.  A faster solution would be
to write the entire gradient node chunk as Base64 or something similar,
but because this is the first place I've had to do something like this,
I may look at optimizing the XML serializer instead.

In the meantime, I've simply reduced the maximum number of supported
gradient nodes, which considerably reduces the performance impact.  This
whole dialog will get some love post-7.0's release to improve some
irritations, but for now, it is at least fully functional.