pdString: a new, lightweight string-builder class
Because PD internally serializes a lot of program data as XML, pdString
includes a number of XML-specific functions that are otherwise slow and
tedious to perform using string functions (e.g. updating tags in-place).

An underlying VB string object is used as its buffer, which allows us to
still use VB string functions where applicable; note that in some
places, however, we specifically swap in API functions to avoid wasting
time on the null-bytes that (likely) occupy the trailing edge of the
buffer.

Class functions are roughly modeled after C#'s StringBuilder; see
https://msdn.microsoft.com/en-us/library/system.text.stringbuilder(v=vs.110).aspx
for details.  (As always, only the bits relevant to PD have been
implemented; I'm not intending this to be an exact match for the .Net
class.)