Effects > Stylize > Antique: new, improved algorithm and UI
With this commit, all Effects menu items now expose full-featured UIs
for the user to play with.  All shitty one-click effects have been
removed and/or fully rewritten, and that menu is now 100% ready for
7.0's release.

(Also, the new antique filter is pretty cool!  Lots of fun settings to
play with, and a very nice end result IMO.)