Metadata engine: ensure initial autosave/Undo data is created
Because metadata parsing is asynchronous, it is possible (even probable
under some circumstances) that the last bits of metadata arrive after
the image load process has already terminated.  This means that the
program's initial Undo/Redo cache (which contains a full backup of the
original image data, in case the image goes missing during editing) will
*not* contain parts of the image's original metadata.

This is obviously not good, so PD now solves the problem by adding
metadata to the initial Undo/Redo cache, if it arrives late.  Enabling
this was a little wonky, but I'll be testing it intensively while fixing
some other metadata issues before release.