Fix potentially severe selection engine bug, when reading Undo data
The bug actually existed in the underlying pdPackager class, and it
happened when a piece of data was compressed, but the compressed size
was identical to the original, uncompressed size. The data would be
written out to file just fine, but a failed check in a subsequent read
function would assume the data was *not* compressed (because the
compressed size was identical to the original size).  This caused it to
skip the decompression step, which resulted in the gibberish compressed
data being treated as original data, which caused a bunch of subsequent
functions to crash and burn.

The reason I discovered this through PD's Selection class is that the
selection class potentially produces very small files (well under 1kb
for simple shapes like rectangular selections), and those sizes are
sometimes too small for compression engines to do anything.  All I was
seeing was Undo/Redo steps that could inexplicably fail, but not
reliably.  (Turns out the only certain floating-point coordinate
combinations may cause the compressed data to come out to a precisely
identical size.)

Anyway, once I finally tracked this one down, I solved the problem
directly inside pdPackager.  If compression results in an
equal-or-larger-sized chunk of data, we reset all chunk flags and write
the original, uncompressed data instead.  This gives us both size and
performance improvements, and it effectively solves this damn bug, too.